+++ sipXportLib/src/os/OsEventMsg.cpp
@@ -28,7 +28,7 @@
 // Constructor
 OsEventMsg::OsEventMsg(const unsigned char subType,
                        const OsQueuedEvent& rEvent,
-                       const int eventData,
+                       const intptr_t eventData,
                        const OsTime& rTimestamp)
 :  OsMsg(OsMsg::OS_EVENT, subType),
    mEventData(eventData),
@@ -92,7 +92,7 @@ int OsEventMsg::getMsgSize(void) const
 
 // Return the event data that was signaled by the notifier task.
 // Always returns OS_SUCCESS.
-OsStatus OsEventMsg::getEventData(int& rEventData) const
+OsStatus OsEventMsg::getEventData(intptr_t& rEventData) const
 {
    rEventData = mEventData;
    return OS_SUCCESS;
@@ -108,7 +108,7 @@ OsStatus OsEventMsg::getTimestamp(OsTime& rTimestamp) const
 
 // Return the user data specified when the OsQueuedEvent was constructed.
 // Always returns OS_SUCCESS.
-OsStatus OsEventMsg::getUserData(int& rUserData) const
+OsStatus OsEventMsg::getUserData(intptr_t& rUserData) const
 {
    rUserData = mUserData;
    return OS_SUCCESS;
