+++ sipXportLib/src/test/os/OsTimerTest.cpp
@@ -195,7 +195,7 @@ public:
        gCallBackCount++ ;
     }
 
-    static void TVCallback(const int userData, const int eventData)
+    static void TVCallback(const intptr_t userData, const intptr_t eventData)
     {
         OsTimerTest* foo = (OsTimerTest*) userData;
         foo->setTVCalled();
@@ -232,7 +232,7 @@ public:
         OsTimer* pTimer ;
         OsStatus returnValue;
         long diffUSecs ; 
-        pNotifier = new OsCallback((int)this, TVCallback);
+        pNotifier = new OsCallback((intptr_t)this, TVCallback);
         pTimer = new OsTimer(*pNotifier);
         gCallBackCount = 0;
         g_get_current_time(&startTV);
@@ -296,7 +296,7 @@ public:
             OsTime timeToWait(testData[i].seconds,
                 testData[i].milliseconds*OsTime::USECS_PER_MSEC) ; 
 
-            pNotifier = new OsCallback((int)this, TVCallback);
+            pNotifier = new OsCallback((intptr_t)this, TVCallback);
             pTimer = new OsTimer(*pNotifier);
 
             expectedWaitUSecs= SecsToUsecs(testData[i].seconds) + 
@@ -344,7 +344,7 @@ public:
         long expectedWaitUSecs;
         long diffUSecs;
 
-        pNotifier = new OsCallback((int)this, TVCallback);
+        pNotifier = new OsCallback((intptr_t)this, TVCallback);
         pTimer = new OsTimer(*pNotifier);
 
         expectedWaitUSecs=(1*OsTime::USECS_PER_SEC) + (250*OsTime::USECS_PER_MSEC) ;  
@@ -379,7 +379,7 @@ public:
         OsTimer* pTimer ;
         UtlBoolean returnValue ;  
         long diffUSecs ; 
-        pNotifier = new OsCallback((int)this, TVCallback);
+        pNotifier = new OsCallback((intptr_t)this, TVCallback);
         pTimer = new OsTimer(*pNotifier);
         // Create an OsDateTime object for 2 seconds in the future
         // and call oneShotAt.
@@ -428,7 +428,7 @@ public:
     {
         OsCallback* pNotifier; 
         OsTimer* pTimer ;
-        pNotifier = new OsCallback((int)this, TVCallback);
+        pNotifier = new OsCallback((intptr_t)this, TVCallback);
         pTimer = new OsTimer(*pNotifier);
         gCallBackCount = 0;
         pTimer->oneshotAfter(oneSecond) ; 
@@ -451,7 +451,7 @@ public:
         OsCallback* pNotifier ;
         OsTimer* pTimer ; 
         UtlBoolean returnValue ; 
-        pNotifier = new OsCallback((int)this, TVCallback);
+        pNotifier = new OsCallback((intptr_t)this, TVCallback);
         pTimer = new OsTimer(*pNotifier);
         gCallBackCount = 0;
         returnValue = pTimer->periodicEvery(twoSeconds, twoSeconds) ; 
@@ -475,7 +475,7 @@ public:
     {
         OsCallback* pNotifier ;
         OsTimer* pTimer ; 
-        pNotifier = new OsCallback((int)this, TVCallback);
+        pNotifier = new OsCallback((intptr_t)this, TVCallback);
         pTimer = new OsTimer(*pNotifier);
         gCallBackCount = 0;
         pTimer->periodicEvery(OsTime::NO_WAIT_TIME, hundredMsec);
@@ -498,7 +498,7 @@ public:
         OsCallback* pNotifier ;
         OsTimer* pTimer ; 
         long diffUSecs ; 
-        pNotifier = new OsCallback((int)this, TVCallback);
+        pNotifier = new OsCallback((intptr_t)this, TVCallback);
         pTimer = new OsTimer(*pNotifier);
         gCallBackCount = 0;
         g_get_current_time(&startTV);
@@ -530,7 +530,7 @@ public:
         OsTimer* pTimer ; 
         OsTimer* pTimer2 ; 
         long diffUSecs ; 
-        pNotifier = new OsCallback((int)this, TVCallback);
+        pNotifier = new OsCallback((intptr_t)this, TVCallback);
         pTimer = new OsTimer(*pNotifier);
         gCallBackCount = 0;
         pTimer->periodicEvery(oneSecond, twoSeconds) ; 
@@ -578,7 +578,7 @@ public:
     {
         OsCallback* pNotifier ;
         OsTimer* pTimer ; 
-        pNotifier = new OsCallback((int)this, TVCallback);
+        pNotifier = new OsCallback((intptr_t)this, TVCallback);
         pTimer = new OsTimer(*pNotifier);
         gCallBackCount = 0;
         pTimer->periodicEvery(oneSecond, twoSeconds) ; 
@@ -597,7 +597,7 @@ public:
     void testStartFire()
     {
        OsStatus returnValue;
-       OsCallback notifier((int) this, TVCallback);
+       OsCallback notifier((intptr_t) this, TVCallback);
        OsTimer timer(notifier);
        gCallBackCount = 0;
        returnValue =  timer.oneshotAfter(oneSecond);
@@ -611,7 +611,7 @@ public:
     void testStartStop()
     {
        OsStatus returnValue;
-       OsCallback notifier((int) this, TVCallback);
+       OsCallback notifier((intptr_t) this, TVCallback);
        OsTimer timer(notifier);
        gCallBackCount = 0;
        returnValue = timer.oneshotAfter(oneSecond);
@@ -628,7 +628,7 @@ public:
     void testStartPeriodicStop()
     {
        OsStatus returnValue;
-       OsCallback notifier((int) this, TVCallback);
+       OsCallback notifier((intptr_t) this, TVCallback);
        OsTimer timer(notifier);
        gCallBackCount = 0;
        returnValue = timer.periodicEvery(oneSecond, twoSeconds);
@@ -649,7 +649,7 @@ public:
     void testStartDelete()
     {
        OsStatus returnValue;
-       OsCallback notifier((int) this, TVCallback);
+       OsCallback notifier((intptr_t) this, TVCallback);
        OsTimer* pTimer = new OsTimer(notifier);
        gCallBackCount = 0;
        returnValue = pTimer->oneshotAfter(oneSecond);
@@ -667,7 +667,7 @@ public:
     void testStartStopStartFire()
     {
        OsStatus returnValue;
-       OsCallback notifier((int) this, TVCallback);
+       OsCallback notifier((intptr_t) this, TVCallback);
        OsTimer timer(notifier);
        gCallBackCount = 0;
        returnValue = timer.oneshotAfter(oneSecond);
@@ -687,7 +687,7 @@ public:
     void testStartStart()
     {
        OsStatus returnValue;
-       OsCallback notifier((int) this, TVCallback);
+       OsCallback notifier((intptr_t) this, TVCallback);
        OsTimer timer(notifier);
        gCallBackCount = 0;
        returnValue = timer.oneshotAfter(oneSecond);
@@ -702,7 +702,7 @@ public:
     void testStartStopStop()
     {
        OsStatus returnValue;
-       OsCallback notifier((int) this, TVCallback);
+       OsCallback notifier((intptr_t) this, TVCallback);
        OsTimer timer(notifier);
        gCallBackCount = 0;
        returnValue = timer.oneshotAfter(oneSecond);
@@ -720,7 +720,7 @@ public:
     void testStartFireStop()
     {
        OsStatus returnValue;
-       OsCallback notifier((int) this, TVCallback);
+       OsCallback notifier((intptr_t) this, TVCallback);
        OsTimer timer(notifier);
        gCallBackCount = 0;
        returnValue = timer.oneshotAfter(oneSecond);
@@ -735,7 +735,7 @@ public:
     void testStop()
     {
        OsStatus returnValue;
-       OsCallback notifier((int) this, TVCallback);
+       OsCallback notifier((intptr_t) this, TVCallback);
        OsTimer timer(notifier);
        gCallBackCount = 0;
        returnValue = timer.stop();
@@ -790,7 +790,7 @@ public:
        // "started" but not put the timer back in the timer queue, so
        // when the update message caused OsTimerTask::removeTimer to run, it
        // couldn't find the timer.
-       OsCallback notifier((int) this, TVCallback);
+       OsCallback notifier((intptr_t)this, TVCallback);
        OsTimer timer(notifier);
 
        // Start timer to fire in 1 sec, and every 1 sec after that.
